+++ ont_fast5_api/conversion_tools/compress_fast5.py
@@ -10,10 +10,11 @@ from ont_fast5_api.conversion_tools.conversion_utils import get_fast5_file_list,
 from ont_fast5_api.fast5_file import Fast5File, EmptyFast5
 from ont_fast5_api.fast5_interface import is_multi_read
 from ont_fast5_api.multi_fast5 import MultiFast5File, copy_attributes
+from ont_fast5_api.static_data import OPTIONAL_READ_GROUPS
 
 
 def compress_batch(input_folder, output_folder, target_compression, recursive=True, threads=1, follow_symlinks=True,
-                   in_place=False):
+                   in_place=False, sanitize=False):
     # We require an absolute input path to we can replicate the data structure relative to it later on
     input_folder = os.path.abspath(input_folder)
 
@@ -39,7 +40,7 @@ def compress_batch(input_folder, output_folder, target_compression, recursive=Tr
             output_path = os.path.join(output_folder, os.path.relpath(input_path, input_folder))
 
         pool.apply_async(func=compress_file,
-                         args=(input_path, output_path, target_compression),
+                         args=(input_path, output_path, target_compression, sanitize),
                          callback=update)
 
     # Tear down the process pool and pbar. We can't use contextmanagers since we need to close() then join()
@@ -48,17 +49,17 @@ def compress_batch(input_folder, output_folder, target_compression, recursive=Tr
     pbar.finish()
 
 
-def compress_file(input_file, output_file, target_compression):
+def compress_file(input_file, output_file, target_compression, sanitize=False):
     try:
         os.makedirs(os.path.dirname(output_file), exist_ok=True)
         if is_multi_read(input_file):
             with MultiFast5File(input_file, 'r') as input_f5, MultiFast5File(output_file, 'a') as output_f5:
                 for read in input_f5.get_reads():
-                    output_f5.add_existing_read(read, target_compression)
+                    output_f5.add_existing_read(read, target_compression, sanitize=sanitize)
         else:
             with Fast5File(input_file, 'r') as input_f5, \
                     EmptyFast5(output_file, 'a') as output_f5:
-                compress_single_read(output_f5, input_f5, target_compression)
+                compress_single_read(output_f5, input_f5, target_compression, sanitize=sanitize)
     except Exception as e:
         # Error raised in Pool.async will be lost so we explicitly print them.
         logging.exception(e)
@@ -66,7 +67,7 @@ def compress_file(input_file, output_file, target_compression):
     return (input_file, output_file)
 
 
-def compress_single_read(output_f5, read_to_copy, target_compression):
+def compress_single_read(output_f5, read_to_copy, target_compression, sanitize=False):
     read_id = read_to_copy.get_read_id()
     raw_dataset_name = read_to_copy.raw_dataset_name
     raw_group_name = read_to_copy.raw_dataset_group_name
@@ -81,6 +82,9 @@ def compress_single_read(output_f5, read_to_copy, target_compression):
         copy_attributes(read_to_copy.handle.attrs, output_f5.handle)
         for subgroup in read_to_copy.handle:
             if subgroup not in raw_dataset_name:
+                if sanitize and subgroup in OPTIONAL_READ_GROUPS:
+                    # skip optional groups when sanitizing
+                    continue
                 output_f5.handle.copy(read_to_copy.handle[subgroup], subgroup)
             else:
                 raw_attrs = read_to_copy.handle[raw_group_name].attrs
@@ -101,6 +105,8 @@ def main():
 
     parser.add_argument('-c', '--compression', required=True, choices=list(COMPRESSION_MAP.keys()),
                         help="Target output compression type")
+    parser.add_argument('--sanitize', action='store_true',
+                        help="Clean output files of optional groups and datasets (e.g. 'Analyses')")
     parser.add_argument('-t', '--threads', type=int, default=1, required=False,
                         help="Maximum number of threads to use")
     parser.add_argument('--recursive', action='store_true',
@@ -112,6 +118,8 @@ def main():
 
     if args.input_path == args.save_path:
         raise ArgumentError(save_arg, "--input_path and --save_path must be different locations, or use --in_place")
+    if args.sanitize and args.save_path is None:
+        raise ArgumentError(save_args, "--save_path must be given if using --sanitize")
 
     compress_batch(input_folder=args.input_path,
                    output_folder=args.save_path,
@@ -119,7 +127,8 @@ def main():
                    threads=args.threads,
                    recursive=args.recursive,
                    follow_symlinks=not args.ignore_symlinks,
-                   in_place=args.in_place)
+                   in_place=args.in_place,
+                   sanitize=args.sanitize)
 
 
 if __name__ == '__main__':